requirements.txt
selenium==4.17.2
flask==2.0.2
json5==0.9.6
gunicorn==20.1.0
werkzeug==2.2.2
google-cloud-storage
google-cloud-bigquery
python-dotenv
google-cloud-pubsub